1400-Druid/Codes/requirements.txt
kafka-python
faker

